Made Small Changes to Starter World
I wanted to increase the pace of the starter world a bit, so I added this new upgrade and adjusted a few other prices slightly to compensate. This should increase the pace of the start of the world a bit (with the hope that it keeps more new players interested). Needs a sprite.