Add Investi-gate-tions sprite
Added sprite for technology "Investi-agate-tions"

Description before unlocking: 
"Temporarily commit 500 investigators to investigating the gate. Uncover its mysteries in record time!"

Description after unlocking:
"Gate investigated. We think we have a way to open it, but it's not gonna be easy... the mechanism is complicated."

History log:
"Using our reverse-engineering techniques, the investigators have come up with a theory on how to open the gate!"