Add sociology sprite
Added sociology .png sprite technology from Volcanic Ocean

Description before unlocking: 
What is a frenzy? Why do we obey our orders? Why do we work together?

Description after unlocking:
Shrimp and sponge farms are twice as effective now that we can grasp the idea of a social system being different than ours.

History log:
Progress, prosperity, organization, order. We choose to be a part of this society, and that is what creates it.