Add civilContact.png sprite
Added .png sprite for frigid Ocean technology Civil Contact

Before unlocking: We see a number of strange structures through the gap in the ice. What are we looking at, exactly?

After unlocking: We visited the structures, and it turns out it's an entire civilization!

Log: Found the squids. They can be enlisted to help catch fish. Also moved to a less frozen place.