Add sprite Apologetic Amnesty
Added the sprite for the Volcanic Ocean's technology "Apologetic Amnesty"

Apologetic Amnesty

Before unlocking: The king is willing to show some mercy. Give him what he wants - repayment for what we stole, apparently.

After unlocking: The king took us into an old, secret room behind the throne. His assistants flipped the right levers and switches, and the gate opened.

Log: The king has decided to open the gate for us, on the condition that he takes control of the frenzy. A deal's a deal.