sustainableSolutions
Add sprite for Volcanic Ocean's "Sustainable Solutions" technology

Before unlocking: As in, find solutions that the shrimp prefer.

After unlocking: Sponge farms are thirty-two times as effective, coral farms are twice as effective. We are diverting a lot of attention to the farms now. Maybe too much? Nah.

Log: So, as it turns out, the shrimp preference is merely to farm for sponge and coral instead of gathering them. That sounds about right.